+++ include/vrv/graphic.h
@@ -48,6 +48,20 @@ public:
     int GetDrawingHeight(int unit, int staffSize) const;
     ///@}
 
+    //----------//
+    // Functors //
+    //----------//
+
+    /**
+     * Interface for class functor visitation
+     */
+    ///@{
+    FunctorCode Accept(Functor &functor) override;
+    FunctorCode Accept(ConstFunctor &functor) const override;
+    FunctorCode AcceptEnd(Functor &functor) override;
+    FunctorCode AcceptEnd(ConstFunctor &functor) const override;
+    ///@}
+
 protected:
     //
 private: