+++ include/vrv/system.h
@@ -155,12 +155,6 @@ public:
     double EstimateJustificationRatio(const Doc *doc) const;
 
     /**
-     * Convert mensural MEI into cast-off (measure) segments looking at the barLine objects.
-     * Segment positions occur where a barLine is set on all staves.
-     */
-    void ConvertToCastOffMensuralSystem(Doc *doc, System *targetSystem);
-
-    /**
      * Reverse of ConvertToCastOffMensural()
      */
     void ConvertToUnCastOffMensuralSystem();