+++ src/setscoredeffunctor.cpp
@@ -230,6 +230,16 @@ FunctorCode ScoreDefSetCurrentFunctor::VisitPage(Page *page)
     return FUNCTOR_CONTINUE;
 }
 
+FunctorCode ScoreDefSetCurrentFunctor::VisitProport(Proport *proport)
+{
+    if (proport->GetType() == "cmme_tempo_change") return FUNCTOR_SIBLINGS;
+    assert(m_currentStaffDef);
+    StaffDef *upcomingStaffDef = m_upcomingScoreDef.GetStaffDef(m_currentStaffDef->GetN());
+    assert(upcomingStaffDef);
+    upcomingStaffDef->SetCurrentProport(proport);
+    return FUNCTOR_CONTINUE;
+}
+
 FunctorCode ScoreDefSetCurrentFunctor::VisitScore(Score *score)
 {
     m_currentScore = score;
@@ -539,12 +549,12 @@ FunctorCode ScoreDefSetGrpSymFunctor::VisitGrpSym(GrpSym *grpSym)
         StaffDef *end = vrv_cast<StaffDef *>(scoreDef->FindDescendantByComparison(&compare, level));
 
         if (!start || !end) {
-            LogWarning("Could not find startid/endid on level %d for <'%s'>", level, grpSym->GetID().c_str());
+            LogWarning("Could not find startid/endid on level %d for '%s'", level, grpSym->GetID().c_str());
             return FUNCTOR_CONTINUE;
         }
 
         if (start->GetParent() != end->GetParent()) {
-            LogWarning("<'%s'> has mismatching parents for startid:<'%s'> and endid:<'%s'>", grpSym->GetID().c_str(),
+            LogWarning("'%s' has mismatching parents for startid:%s and endid:%s", grpSym->GetID().c_str(),
                 startId.c_str(), endId.c_str());
             return FUNCTOR_CONTINUE;
         }